+++ Package.swift
@@ -17,6 +17,7 @@ let package = Package(
         .package(url: "https://github.com/sparkle-project/Sparkle", from: "2.8.1"),
         .package(url: "https://github.com/orchetect/MenuBarExtraAccess", exact: "1.2.2"),
         .package(url: "https://github.com/apple/swift-algorithms", from: "1.2.1"),
+        .package(url: "https://github.com/apple/swift-log", from: "1.8.0"),
         .package(url: "https://github.com/openid/AppAuth-iOS", from: "2.0.0"),
         .package(url: "https://github.com/apollographql/apollo-ios", from: "2.0.3"),
         .package(url: "https://github.com/onevcat/Kingfisher", from: "8.6.0"),
@@ -27,6 +28,7 @@ let package = Package(
             name: "RepoBarCore",
             dependencies: [
                 .product(name: "Apollo", package: "apollo-ios"),
+                .product(name: "Logging", package: "swift-log"),
             ],
             swiftSettings: [
                 .enableUpcomingFeature("StrictConcurrency"),
@@ -40,6 +42,7 @@ let package = Package(
                 .product(name: "Algorithms", package: "swift-algorithms"),
                 .product(name: "AppAuth", package: "AppAuth-iOS"),
                 .product(name: "Kingfisher", package: "Kingfisher"),
+                .product(name: "Logging", package: "swift-log"),
             ],
             exclude: ["Resources/Info.plist"],
             swiftSettings: [